requirements.txt
numpy
opencv-python==4.1.1
requests
pyaml
six
boto3
matplotlib